fix examples/dwg2svg2: get the center
* examples/dwg2svg2.c: circle.center was never initialized.
  fix the paperspace limits calculation: LIM not EXT.
  use double, not float.
* examples/dwg_ps.c: use double, not float.

+++ examples/dwg_ps.c
@@ -28,11 +28,11 @@
 void
 create_postscript(Dwg_Data *dwg, char *output)
 {
-  float dx;
-  float dy;
-  float scale_x;
-  float scale_y;
-  float scale;
+  double dx;
+  double dy;
+  double scale_x;
+  double scale_y;
+  double scale;
   long unsigned i;
   //FILE *fh;
   PSDoc *ps;
